requirements.txt
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
openai>=1.3.0
pyaudio>=0.2.11
aiohttp>=3.8.0
python-dotenv>=1.0.0
